studentsdb/requirements.txt
Django==1.7.1
Pillow==2.7.0
MySQL-python==1.2.5
django-crispy-forms==1.4.0
django-contact-form==1.0
django-simple-captcha==0.4.4